+++ R/proportional_hazards_data.R
@@ -1,5 +1,5 @@
 
-# parnip just contains the model specification, the engines are the censored package.
+# parsnip just contains the model specification, the engines are the censored package.
 
 set_new_model("proportional_hazards")
 set_model_mode("proportional_hazards", "censored regression")